OF-2152: Apply more explicit address verification
This commit takes the explicit 'from' address verification that was put in place for Presence broadcast handling, and applies the same to Message broadcast handling.

The check has been moved up a bit, and has been applied at the constructor of the tasks that are executed asynchronously. These can be expected to generate more useful stack traces.